+++ src/Commands/Moderation/Unmute.js
@@ -23,12 +23,12 @@ run(message) {
     const role = message.guild.roles.cache.find(role => role.name === 'Muted');
 
     if (muted.hasPermission(['KICK_MEMBERS', 'BAN_MEMBERS'])) return message.reply('  :warning: This member cant be unmuted.');
-
+    if (muted.voice.channel) return muted.voice.setMute(false)
     message.delete()
     muted.roles.remove(role)
     message.channel.send(`*${muted.user.username}* was successfully unmuted.`)
     muted.send(`Hello, you have been unmuted in **${message.guild.name}** by: **${message.author.tag}**`).catch(err => console.log(err))
-    muted.voice.setMute(false)
+  
 
     const embed = new MessageEmbed()
                 .setColor('RED')
